refactor(rebalancer): move factory into a factories directory
