+++ typescript/cli/src/rebalancer/factories/RebalancerContextFactory.ts
@@ -7,9 +7,9 @@ import {
 } from '@hyperlane-xyz/sdk';
 import { objMap, objMerge } from '@hyperlane-xyz/utils';
 
-import { Metrics } from './metrics/Metrics.js';
-import { PriceGetter } from './metrics/PriceGetter.js';
-import { Monitor } from './monitor/Monitor.js';
+import { Metrics } from '../metrics/Metrics.js';
+import { PriceGetter } from '../metrics/PriceGetter.js';
+import { Monitor } from '../monitor/Monitor.js';
 
 export class RebalancerContextFactory {
   /**
